move the copy utility function in src/utils.ts
